refactor(graft/evm/sync): reduce the diff
