+++ graft/evm/sync/leaf/BUILD.bazel
@@ -20,7 +20,10 @@ go_library(
 
 graft_go_test(
     name = "leaf_test",
-    srcs = ["syncer_test.go"],
+    srcs = [
+        "keys_test.go",
+        "syncer_test.go",
+    ],
     embed = [":leaf"],
     deps = [
         "//graft/evm/sync/synctest",
