+++ graft/evm/sync/synctest/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "synctest",
     testonly = True,
     srcs = [
-        "accounts.go",
         "blocks.go",
         "fetchers.go",
         "state.go",
